add QFN-21_4x4_P0.4mm_SC8721 package.
This package is firstly auto-generated from QFN32 wizard. The size and offset of small pads are set to the same value
in the QFN32 4x4mm package in official library. The size and offset of large pads are enlarged for large current.
The package name is suffixed with part name, for there are five unusual pins and a 0175mm offset in vertical direction,
which is highly unlikely to be reused for other parts.